pyproject.toml
[project]
name = "adk-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.12.0",
    "opencv-python>=4.12.0.88",
]
